StringCalculator와 main, input 클래스 분리
